requirements.txt
# 量化交易系统依赖

# ── OKX SDK（含 requests / aiohttp 等传递依赖）──────────────────
python-okx>=0.3.0

# ── 异步 WebSocket ─────────────────────────────────────────────
websockets>=12.0

# ── 配置文件解析 ────────────────────────────────────────────────
PyYAML>=6.0

# ── 回测 HTML 报告（交互式图表）────────────────────────────────
plotly>=5.18.0

# ── 测试框架 ────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0

# ── 预留：Interactive Brokers ──────────────────────────────────
# ibapi>=10.19.1       # IB 官方 Python API（需要从 IB 官网下载）
# ib_insync>=0.9.86    # 社区封装

# ── 预留：CTP 期货 ──────────────────────────────────────────────
# vnpy_ctp>=3.7.0      # 需要特定平台，仅 Win/Linux
